Rename directive module to directives
